requirements.txt
appdirs==1.4.4
autopep8==1.5.7
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-vision==2.3.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
huggingface-hub==0.0.12
idna==2.10
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.0
openpyxl==3.0.7
packaging==21.0
pathspec==0.9.0
pandas==1.3.0
pdf2image==1.16.0
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.3.1
pkg-resources==0.0.0
poppler-utils==0.1.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.10.1
pyparsing==2.4.7
python-dotenv==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==1.0.4
typed-ast==1.4.3
sklearn==0.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Wand==0.6.6
Werkzeug==2.0.1
zipp==3.5.0

